outsourced LeptonSFHelper class to make C code more readable
